-BIRG REFACTORING- -->Separated everything in 4 concepts ---->Engine (apply procedures) ---->Services (serialize/deserialize) ---->Model (classes) ---->Screens (display engine results)
Engine asks services to pack and unpack data, keeps everything ins a "GameState.cs", when a screen requires it, Engine also does stuffs and gives results that can be displayed, as well as saving said results to the GameState. Engine uses the contents of Model, to give screens the results.

**Created a mission validation method, any further changes to mission validation should be done on that method. It checks wether all the required parameters were filled by the user or no.

**Started working on a "Calendar.cs", to keep track of time.

**Added time buttons on the main menu.